PMM-10579 refactoring
