+++ public/app/percona/shared/core/reducers/services/index.ts
@@ -1,0 +1,6 @@
+import servicesReducer from './services';
+export * from './services';
+export * from './services.types';
+export * from './services.utils';
+
+export default servicesReducer;
